[oauth] Move oauth (back) to common.
We moved oauth from internal/oauth to common/oauth in #592,
but users need access to oauth protos to programmatically generate
Cloudprober configs.
