+++ internal/rds/client/proto/config.proto
@@ -9,7 +9,7 @@ syntax = "proto2";
 
 package cloudprober.rds;
 
-import "github.com/cloudprober/cloudprober/internal/oauth/proto/config.proto";
+import "github.com/cloudprober/cloudprober/common/oauth/proto/config.proto";
 import "github.com/cloudprober/cloudprober/common/tlsconfig/proto/config.proto";
 import "github.com/cloudprober/cloudprober/internal/rds/proto/rds.proto";
