+++ probes/browser/storage/storage_gcs.go
@@ -21,8 +21,8 @@ import (
 	"net/http"
 	"path"
 
-	"github.com/cloudprober/cloudprober/internal/oauth"
-	oauthconfigpb "github.com/cloudprober/cloudprober/internal/oauth/proto"
+	"github.com/cloudprober/cloudprober/common/oauth"
+	oauthconfigpb "github.com/cloudprober/cloudprober/common/oauth/proto"
 	"github.com/cloudprober/cloudprober/logger"
 	configpb "github.com/cloudprober/cloudprober/probes/browser/proto"
 	"golang.org/x/oauth2"
