+++ probes/grpc/proto/config.proto
@@ -2,7 +2,7 @@ syntax = "proto2";
 
 package cloudprober.probes.grpc;
 
-import "github.com/cloudprober/cloudprober/internal/oauth/proto/config.proto";
+import "github.com/cloudprober/cloudprober/common/oauth/proto/config.proto";
 import "github.com/cloudprober/cloudprober/common/tlsconfig/proto/config.proto";
 
 option go_package = "github.com/cloudprober/cloudprober/probes/grpc/proto";
