+++ probes/http/proto/config.proto
@@ -2,7 +2,7 @@ syntax = "proto2";
 
 package cloudprober.probes.http;
 
-import "github.com/cloudprober/cloudprober/internal/oauth/proto/config.proto";
+import "github.com/cloudprober/cloudprober/common/oauth/proto/config.proto";
 import "github.com/cloudprober/cloudprober/common/tlsconfig/proto/config.proto";
 import "github.com/cloudprober/cloudprober/metrics/payload/proto/config.proto";
